fix errors

+++ src/main/java/com/klip/petcare/service/core/OwnerService.java
@@ -3,5 +3,5 @@ package com.klip.petcare.service.core;
 import com.klip.petcare.dto.request.OwnerRequestDTO;
 import com.klip.petcare.service.base.GenericService;
 
-public interface OwnerService extends GenericService<OwnerRequestDTO, Long> {
+public interface OwnerService extends GenericService<OwnerRequestDTO, Integer> {
 }
